Convert unit tests to typescript.
