+++ custom_typings/yeoman-assert.d.ts
@@ -1,0 +1,5 @@
+declare module 'yeoman-assert' {
+  export function file(filenames: string[]);
+  export function fileContent(path: string, content: string);
+  export function noFile(filenames: string[]);
+}
