+++ package.json
@@ -84,7 +84,10 @@
   "devDependencies": {
     "@polymer/tools-common": "^1.0.1",
     "@types/fs-extra": "0.0.37",
+    "@types/mocha": "^2.2.44",
     "@types/node": "=6.0.65",
+    "@types/sinon": "^4.0.0",
+    "@types/yeoman-test": "^1.7.3",
     "babili": "^0.0.12",
     "chai": "^3.5.0",
     "clang-format": "=1.0.50",
@@ -99,7 +102,7 @@
     "memory-streams": "^0.1.0",
     "mocha": "^3.0",
     "run-sequence": "^1.2.0",
-    "sinon": "^1.17.4",
+    "sinon": "^1.17.7",
     "tmp": "0.0.31",
     "tslint": "^4.0.2",
     "typescript": "^2.2.0",
